Fix the M5 CI failures: restart-send UAF and release -Werrors
Two distinct classes of failure in run 33186812802:

1. heap-use-after-free (asan+tsan, heyaki_m4_shutdown_matrix):
   send_restart_frame can fail the superseded session synchronously; the
   close observer retires (and erases) the owning attempt while the send
   path is still dereferencing it. PeerSession public entry points now
   hold a strong self reference for the duration of the call, and the
   Node restart senders copy the session pointer before sending and stop
   touching the attempt afterwards. Verified: shutdown matrix clean over
   repeated asan and tsan (setarch -R) runs, restart e2e and M5 session
   suites tsan-clean.

2. release -Werror builds failed to compile:
   - Frame designated initializer omitted message_id; use plain member
     assignment in send_stream_frame.
   - parse_stream_id was unused after the Protobuf stream-id fix; route
     handle_data's id extraction through it so both layouts share the
     zero/shape checks.
   - M5 config fields appended to VerifiedPeerSessionConfig and
     PairingServiceConfig lacked default member initializers, tripping
     -Wmissing-field-initializers at every pre-existing positional call
     site (node.cpp restart path, m4_peer_session_test, pairing test);
     TrustGrantRecord::expires_unix_milliseconds likewise. All get DMIs
     so existing callers stay valid; the restart-path config lists every
     field explicitly.
   - drop now-unused helpers (physical_kind_for_frame, random_grant_id),
   unused lambda/bind parameters, unused run_stream_view/run_command
   parameters, and a sign-conversion in the scheduler benchmark.
   - m5_pairing_service_test disables the OS secret backend like the
     other profile tests (glib allocator noise under tsan) and sets
     owner-only directory permissions.

+++ tests/performance/m5_scheduler_benchmark.cpp
@@ -75,7 +75,7 @@ int main() {
       if (!next.has_value()) break;
       if (next->frame_class == FrameClass::control) {
         control_sent += 1U;
-        control_position_sum += slot;
+        control_position_sum += static_cast<std::size_t>(slot);
         control_observations += 1U;
         max_control_deficit = std::max(max_control_deficit, bulk_sent - bulk_before);
         bulk_before = bulk_sent;
